Find stabilised small world networks

Compares small world networks in which variation in component response rate does not vary to random matrices in which this variation for 2 to max_sp components. This function is to be used for small world networks, which are constructed using the method of Watts and Strogatz (Nature vol. 393, 1998). Off-diagonal elements have a mean of 'mn' and a standard deviation of 'sigma'.

rand_gen_swn(max_sp, iters, int_type = 0, rmx = 0.4, C = 1, by = 4,
  sigma = 0.4, Kdiv = 2, beta = 0.2, mn = 0, dval = 1, g_dist = 1,
  g_mn = 1, g_sd = 1)

Arguments

max_sp
Maximum number of components to randomise

iters
Number of iterations (i.e., random matrices) per component

int_type
Type of interaction between components including random (0), competitor (1), mutualist (2), predator-prey (3), and cascade model (4)

rmx
Standard deviation of population growth rates (for feasibility)

C
Connectedness of matrices (i.e., probability of non-zero matrix element components.

by
Sequence between component number to randomise (e.g., '2': 2, 4, 6)

sigma
Standard deviation of interaction strength among network elements

Kdiv
Value to divide the component number by to produce the parameter K for creating the small world network. For example, if S = 32 and K = 8, then the small world network will be created from a regular network in which each component is connected to 32/8 = 4 other components. This needs to be used cautiously to avoid generating non-even values of K.

beta
Probability that a random interaction in a regular network is rewired (parameter p in Watts and Strogatz 1998)

mn
Mean interaction strength among network elements

dval
Self-regulation of network elements (1 by default)

Value

A table of stability results, where rows summarise for each component number (S) the number of stable or unstable (also, feasible and infeasible) random matrices produced.

Examples

rand_gen_swn(max_sp = 16, iters = 4);

#> [1] 4
#> [1] 8
#> [1] 12
#> [1] 16

#>       S A_unstable A_stable M_unstable M_stable A_stabilised A_destabilised
#> [1,]  4          0        4          0        4            0              0
#> [2,]  8          0        4          0        4            0              0
#> [3,] 12          2        2          2        2            0              0
#> [4,] 16          2        2          2        2            0              0
#>      A_infeasible A_feasible M_infeasible M_feasible A_made_feasible
#> [1,]            4          0            4          0               0
#> [2,]            3          1            4          0               0
#> [3,]            4          0            4          0               0
#> [4,]            4          0            4          0               0
#>      A_made_infeasible       A_rho      M_rho    rho_diff     rho_abs complex_A
#> [1,]                 0  0.34651869  0.2553346 -0.09118413 -0.05152100 0.4688756
#> [2,]                 1 -0.14611317 -0.1908964 -0.04478323  0.06940355 0.6369623
#> [3,]                 0  0.02379225  0.1034654  0.07967315 -0.01074880 0.8260054
#> [4,]                 0 -0.01339716  0.0698931  0.08329027  0.08267907 1.0928594
#>      complex_M       A_eig        M_eig       LR_A       UR_A        LR_M
#> [1,] 0.4947858 -0.37206920 -0.102299753 -0.5618291 -0.2323655 -0.23037739
#> [2,] 0.7501027 -0.33730250 -0.046182904 -0.4973343 -0.2046534 -0.08195856
#> [3,] 0.9247797 -0.05740832 -0.004721538 -0.2691700  0.2040471 -0.11379478
#> [4,] 1.2415699  0.04861164  0.146332746 -0.1691847  0.3179063 -0.03433567
#>             UR_M         C
#> [1,] -0.03665147 1.0000000
#> [2,] -0.02926044 0.6517857
#> [3,]  0.14152773 0.6060606
#> [4,]  0.39637017 0.5604167