Find a stabilised system for correlated systems

Compares random matrices in which variation in component response rate does not vary to random matrices in which this variation for a fixed number of S components. This function is to be used for random networks, in which off-diagonal elements are randomly sampled from a normal distribution with a mean of 'mn', a standard deviation of 'sigma', and a pre-specified set of correlation structures.

rand_rho_var(S, rhos, iters, int_type = 0, rmx = 0.4, C = 1,
  sigma = 0.4, mn = 0, dval = 1)

Arguments

S
Number of components to randomise

rhos
Vector of correlations to simulate

iters
Number of iterations (i.e., random matrices) per component

rmx
Standard deviation of population growth rates (for feasibility)

C
Connectedness of matrices (i.e., probability of non-zero matrix element components.

sigma
Standard deviation of interaction strength among network elements

mn
Mean interaction strength among network elements

dval
Self-regulation of network elements (1 by default)

Value

A table of stability results, where rows summarise for each correlation (rho) the number of stable or unstable (also, feasible and infeasible) random matrices produced.

Examples

rand_rho_var(S = 10, rhos = c(-0.2, 0, 0.2), iters = 4);

#> [1] -0.2
#> [1] 0
#> [1] 0.2

#>      rho_set A_unstable A_stable M_unstable M_stable A_stabilised
#> [1,]    -0.2          0        4          0        4            0
#> [2,]     0.0          1        3          1        3            0
#> [3,]     0.2          4        0          4        0            0
#>      A_destabilised A_infeasible A_feasible M_infeasible M_feasible
#> [1,]              0            4          0            4          0
#> [2,]              0            4          0            4          0
#> [3,]              0            4          0            4          0
#>      A_made_feasible A_made_infeasible         A_rho       M_rho    rho_diff
#> [1,]               0                 0 -2.000000e-01 -0.13767683  0.06232317
#> [2,]               0                 0 -1.721152e-17 -0.05092841 -0.05092841
#> [3,]               0                 0  2.000000e-01  0.09336301 -0.10663699
#>          rho_abs complex_A complex_M       A_eig       M_eig        LR_A
#> [1,] -0.06232317  1.341245  1.519175 -0.38051454 -0.21545315 -0.67393031
#> [2,]  0.07874301  1.241896  1.378136  0.07366909  0.01682414 -0.09747301
#> [3,] -0.10663699  1.379337  1.574762  0.42074448  0.41254535  0.17018199
#>            UR_A        LR_M       UR_M
#> [1,] -0.1130230 -0.35972748 -0.0846676
#> [2,]  0.5352256 -0.13420338  0.2625266
#> [3,]  1.0506072  0.07825166  1.1795419