Find stable bimodal systems

Produces random matrices and determines whether or not they are stable using eigenanalysis given two different component response rates of 1.95 and 0.05 over some number of iterations

stab_bgamma(S = 200, C = 0.05, Osd = 0.4, iters = 10000)

Arguments

S
Size of the complex system

C
Connectedness of complex system

Osd
Standard deviation of interaction strength

iters
Number of random matrices to simulate

Value

A table showing whether matrices are unstable (0) or stable (1) before (column 1) and after (column 2) including variation in component response rate in eigenanalysis for some number (rows) of random matrices.

Examples

dat <- stab_bgamma(iters = 4);

#> [1] 4 0 0
#> [1] 3 0 0
#> [1] 2 0 0
#> [1] 1 0 0