Find stabilised scale-free world networks

Compares scale-free networks in which variation in component response rate does not vary to random matrices in which this variation for 'm' to 'max_sp' components. This function is to be used for scale-free networks, which are constructed using the method of Albert and Barab\'asi (Reviews of Modern Physics, Vo. 74, 2002). Off-diagonal elements have a mean of 'mn' and a standard deviation of 'sigma'.

rand_gen_sfn(S, m_vals, iters, int_type = 0, rmx = 0.4, sigma = 0.4,
  mn = 0, dval = 1, g_dist = 1, g_mn = 1, g_sd = 1)

Arguments

S
Number of components to randomise

m_vals
The number of vertices that newly added components will have in the scale free network (increasing this, but not S, increases the complexity by increasing C).

iters
Number of iterations (i.e., random matrices) per component

int_type
Type of interaction between components including random (0), competitor (1), mutualist (2), predator-prey (3), and cascade model (4)

rmx
Standard deviation of population growth rates (for feasibility)

sigma
Standard deviation of interaction strength among network elements

mn
Mean interaction strength among network elements

dval
Self-regulation of network elements (1 by default)

C
Connectedness of matrices (i.e., probability of non-zero matrix element components.

m
The number of vertices that newly added components will have in scale-free network construction

Value

A table of stability results, where rows summarise for each component number (S) the number of stable or unstable (also, feasible and infeasible) random matrices produced.

Examples

rand_gen_sfn(S = 24, m_vals = c(2, 4), iters = 4);

#> [1] 2
#> [1] 4

#>      S A_unstable A_stable M_unstable M_stable A_stabilised A_destabilised
#> [1,] 2          0        4          0        4            0              0
#> [2,] 4          2        2          2        2            0              0
#>      A_infeasible A_feasible M_infeasible M_feasible A_made_feasible
#> [1,]            4          0            4          0               0
#> [2,]            4          0            4          0               0
#>      A_made_infeasible       A_rho       M_rho    rho_diff     rho_abs
#> [1,]                 0 -0.04580611 -0.02512778  0.02067833 -0.01217005
#> [2,]                 0 -0.06617231 -0.09111299 -0.02494068  0.06270303
#>      complex_A complex_M       A_eig       M_eig       LR_A       UR_A
#> [1,] 0.8904277 0.9753995 -0.44171862 -0.09445092 -0.6084554 -0.2584511
#> [2,] 1.1186440 1.3316457  0.03644578  0.10779208 -0.1298998  0.3617342
#>             LR_M        UR_M         C
#> [1,] -0.22631126 -0.01915051 0.1648551
#> [2,] -0.02932164  0.46835811 0.3134058