Find stabilised small world networks

Compares small world networks in which variation in component response rate does not vary to random matrices in which this variation for 2 to max_sp components. This function is to be used for small world networks, which are constructed using the method of Watts and Strogatz (Nature vol. 393, 1998). Off-diagonal elements have a mean of 'mn' and a standard deviation of 'sigma'.

rand_gen_swn(max_sp, iters, int_type = 0, rmx = 0.4, C = 1, by = 4,
  sigma = 0.4, Kdiv = 2, beta = 0.2, mn = 0, dval = 1, g_dist = 1,
  g_mn = 1, g_sd = 1)

Arguments

max_sp
Maximum number of components to randomise

iters
Number of iterations (i.e., random matrices) per component

int_type
Type of interaction between components including random (0), competitor (1), mutualist (2), predator-prey (3), and cascade model (4)

rmx
Standard deviation of population growth rates (for feasibility)

C
Connectedness of matrices (i.e., probability of non-zero matrix element components.

by
Sequence between component number to randomise (e.g., '2': 2, 4, 6)

sigma
Standard deviation of interaction strength among network elements

Kdiv
Value to divide the component number by to produce the parameter K for creating the small world network. For example, if S = 32 and K = 8, then the small world network will be created from a regular network in which each component is connected to 32/8 = 4 other components. This needs to be used cautiously to avoid generating non-even values of K.

beta
Probability that a random interaction in a regular network is rewired (parameter p in Watts and Strogatz 1998)

mn
Mean interaction strength among network elements

dval
Self-regulation of network elements (1 by default)

Value

A table of stability results, where rows summarise for each component number (S) the number of stable or unstable (also, feasible and infeasible) random matrices produced.

Examples

rand_gen_swn(max_sp = 16, iters = 4);

#> [1] 4
#> [1] 8
#> [1] 12
#> [1] 16

#>       S A_unstable A_stable M_unstable M_stable A_stabilised A_destabilised
#> [1,]  4          0        4          0        4            0              0
#> [2,]  8          0        4          0        4            0              0
#> [3,] 12          1        3          1        3            0              0
#> [4,] 16          2        2          2        2            0              0
#>      A_infeasible A_feasible M_infeasible M_feasible A_made_feasible
#> [1,]            3          1            2          2               1
#> [2,]            4          0            4          0               0
#> [3,]            4          0            4          0               0
#> [4,]            4          0            4          0               0
#>      A_made_infeasible       A_rho       M_rho     rho_diff      rho_abs
#> [1,]                 0  0.48002817  0.47099459 -0.009033579 -0.093580519
#> [2,]                 0 -0.20671475 -0.08362409  0.123090655 -0.111340582
#> [3,]                 0  0.01890760 -0.05110404 -0.070011641  0.007472452
#> [4,]                 0 -0.05063846 -0.01701563  0.033622827  0.012830877
#>      complex_A complex_M      A_eig       M_eig       LR_A        UR_A
#> [1,] 0.2604540 0.2967713 -0.6099816 -0.18354042 -0.7520473 -0.27942635
#> [2,] 0.9894985 1.0599191 -0.6602999 -0.23753169 -0.9933007 -0.36713952
#> [3,] 1.0471728 1.1901489 -0.1494897 -0.10419012 -0.5212001  0.01127894
#> [4,] 1.1475764 1.2813031 -0.1131102 -0.04551905 -0.3586204  0.14837376
#>            LR_M         UR_M         C
#> [1,] -0.4752191 -0.027533847 0.9166667
#> [2,] -0.4184222 -0.033937879 0.6696429
#> [3,] -0.3801035  0.002562937 0.5795455
#> [4,] -0.1967054  0.150238150 0.5666667